requirements.txt
google-cloud-aiplatform
google-adk
fastapi
uvicorn
google-cloud-firestore
python-dotenv
pydantic
google-genai
